Note use of pthreads at compile time.
The -lpthread flag is needed in /tools/ to avoid this bug:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=55394

Use -pthread on compile is a good idea, though there are some contexts
where it may be a no-op.  More specifically, the flag -pthread is
equivalent to -lpthread -D_REENTRANT (and it used to define a
different macro).  The _REENTRANT macro (is somewhat questionably)
used to conditionally compile things for threading.  It is commonly
used as an optimization to remove synchronization code from single
threaded translation units.  None of which might matter here, but I
believe it is a no-op when it doesn't matter.

Part of #9 (https://github.com/google/cctz/issues/9).